+++ openidconnect/lib/src/config/openidconfiguration.dart
@@ -11,8 +11,8 @@ class OpenIdConfiguration {
   final String? registrationEndpoint;
   final String? mfaChallengeEndpoint;
 
-  final List<String> scopesSupported;
-  final List<String> claimsSupported;
+  final List<String>? scopesSupported;
+  final List<String>? claimsSupported;
   final List<String>? grantTypesSupported;
   final List<String> responseTypesSupported;
   final List<String> responseModesSupported;
@@ -38,8 +38,8 @@ class OpenIdConfiguration {
     this.revocationEndpoint,
     this.registrationEndpoint,
     this.mfaChallengeEndpoint,
-    required this.scopesSupported,
-    required this.claimsSupported,
+    this.scopesSupported,
+    this.claimsSupported,
     this.grantTypesSupported,
     required this.responseTypesSupported,
     required this.responseModesSupported,
@@ -68,10 +68,12 @@ class OpenIdConfiguration {
         mfaChallengeEndpoint: json["mfa_challenge_endpoint"]?.toString(),
         deviceAuthorizationEndpoint:
             json["device_authorization_endpoint"]?.toString(),
-        scopesSupported:
-            List<String>.from(json["scopes_supported"] as List<dynamic>),
-        claimsSupported:
-            List<String>.from(json["claims_supported"] as List<dynamic>),
+        scopesSupported: json["scopes_supported"] == null
+            ? null
+            : List<String>.from(json["scopes_supported"] as List<dynamic>),
+        claimsSupported: json["claims_supported"] == null
+            ? null
+            : List<String>.from(json["claims_supported"] as List<dynamic>),
         grantTypesSupported: json["grant_types_supported"] == null
             ? null
             : List<String>.from(json["grant_types_supported"] as List<dynamic>),