+++ openidconnect/lib/src/models/requests/password_authorization_request.dart
@@ -2,25 +2,21 @@ part of openidconnect;
 
 class PasswordAuthorizationRequest extends TokenRequest {
   PasswordAuthorizationRequest({
-    required String clientId,
-    String? clientSecret,
-    required Iterable<String> scopes,
+    required super.clientId,
+    super.clientSecret,
+    required super.scopes,
     required String userName,
     required String password,
-    required OpenIdConfiguration configuration,
+    required super.configuration,
     required bool autoRefresh,
-    Iterable<String>? prompts,
+    super.prompts,
     Map<String, String>? additionalParameters,
   }) : super(
-            configuration: configuration,
-            clientId: clientId,
-            clientSecret: clientSecret,
-            grantType: "password",
-            scopes: scopes,
-            prompts: prompts,
-            additionalParameters: {
-              "username": userName,
-              "password": password,
-              ...(additionalParameters ?? {})
-            });
+          grantType: "password",
+          additionalParameters: {
+            "username": userName,
+            "password": password,
+            ...?additionalParameters,
+          },
+        );
 }