+++ openidconnect/lib/src/openidconnect_client.dart
@@ -38,7 +38,6 @@ class OpenIdConnectClient {
     this.audiences,
   });
 
-  @mustCallSuper
   static Future<OpenIdConnectClient> create({
     required String discoveryDocumentUrl,
     required String clientId,
@@ -358,8 +357,8 @@ class OpenIdConnectClient {
       await _completeLogin(response);
 
       if (autoRefresh) {
-        final refreshTime = _identity!.expiresAt
-            .difference(DateTime.now().subtract(Duration(minutes: 1)));
+        var refreshTime = _identity!.expiresAt.difference(DateTime.now().toUtc());
+        refreshTime -= Duration(minutes: 1);
 
         _autoRenewTimer = Future.delayed(refreshTime, refresh);
       }