+++ openidconnect_platform_interface/lib/src/exceptions/logout_exception.dart
@@ -1,4 +1,4 @@
-part of openidconnect;
+part of openidconnect_platform_interface;
 
 class LogoutException extends OpenIdConnectException {
   LogoutException([String? errorMessage]) : super(errorMessage);