+++ openidconnect_platform_interface/lib/src/exceptions/user_info_exception.dart
@@ -1,4 +1,4 @@
-part of openidconnect;
+part of openidconnect_platform_interface;
 
 class UserInfoException extends OpenIdConnectException {
   UserInfoException([String? errorMessage]) : super(errorMessage);