+++ openidconnect_platform_interface/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [1.0.13] - November 10th, 2022
+
+- Update dependencies
+
+## [1.0.11] - December 9th, 2021
+
+- Add Build context to the authorizeInteractive method
+
+## [1.0.4] - November 29th, 2021
+
+- Update dependencies
+
 ## [1.0.3] - September 10th, 2021
 
 - Update to Dart 2.14